git-svn-id: http://svn.mythtv.org/svn/trunk@16245 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmyth/mythwidgets.cpp
@@ -1573,9 +1573,9 @@ void MythListBox::keyPressEvent(QKeyEvent* e)
             else if (action == "NEXTVIEW")
             {
                 int nextItem = currentItem();
-                if (nextItem < count() - 1)
+                if (nextItem < (int)count() - 1)
                     nextItem++;
-                while (nextItem < count() - 1 && text(nextItem)[0] == ' ')
+                while (nextItem < (int)count() - 1 && text(nextItem)[0] == ' ')
                     nextItem++;
                 if (!itemVisible(nextItem))
                     setTopItem(nextItem);
